requirements.txt
fastapi==0.116.1
strawberry-graphql[fastapi]==0.278.0
asyncpg==0.30.0
uvicorn[standard]==0.35.0
python-dotenv==1.1.1